app/database.py
import os
from sqlmodel import create_engine, SQLModel
from dotenv import load_dotenv

# Cargar variables de entorno desde el archivo .env
load_dotenv()

# Obtener variables de entorno
username = os.getenv('DB_USERNAME')
password = os.getenv('DB_PASSWORD')
host = os.getenv('DB_HOST')
database = os.getenv('DB_DATABASE')
port = os.getenv('DB_PORT')

# Crear el engine de SQLModel
DATABASE_URL = f'mysql+mysqlconnector://{username}:{password}@{host}:{port}/{database}'
engine = create_engine(DATABASE_URL)

def create_db_and_tables():
    SQLModel.metadata.create_all(engine)