requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
sqlmodel>=0.0.8
pydantic>=2.3.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
mysql-connector-python>=8.1.0
bcrypt>=4.0.1
python-dotenv>=1.0.0